intelligent-form-agent/requirements.txt
langchain>=0.0.350
langchain-community>=0.0.10

# Local Ollama client - runs models locally without API keys or quotas
ollama>=0.0.3

# PDF Processing
pypdf>=3.15.0

# Vector Database
chromadb>=0.4.22

# Embeddings & models
sentence-transformers>=2.2.2
transformers>=4.30.0

# Text processing
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.7.1

# UI
streamlit>=1.28.0

# Progress
tqdm>=4.66.1
